requirements.txt
python-dotenv==1.0.0
ccxt==4.1.13
pandas==2.0.3
numpy==1.26.2
python-telegram-bot==20.7
ta==0.10.2
aiohttp>=3.9.0
openai>=1.3.7
pytest==7.4.3
pytest-asyncio==0.21.1